Moved body-related types to body package.
