+++ rawhttp-core/src/main/java/rawhttp/core/EagerHttpResponse.java
@@ -3,6 +3,9 @@ package rawhttp.core;
 import java.io.IOException;
 import java.util.Optional;
 import javax.annotation.Nullable;
+import rawhttp.core.body.BodyReader;
+import rawhttp.core.body.ChunkedBodyContents;
+import rawhttp.core.body.EagerBodyReader;
 
 import static rawhttp.core.RawHttpHeaders.Builder.emptyRawHttpHeaders;
