+++ rawhttp-core/src/main/java/rawhttp/core/HttpMessage.java
@@ -5,6 +5,7 @@ import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Optional;
 import javax.annotation.Nullable;
+import rawhttp.core.body.BodyReader;
 import rawhttp.core.body.HttpMessageBody;
 
 /**
