+++ rawhttp-core/src/main/java/rawhttp/core/IOConsumer.java
@@ -2,7 +2,12 @@ package rawhttp.core;
 
 import java.io.IOException;
 
+/**
+ * A Consumer that can throw {@link IOException}.
+ *
+ * @param <T> the argument type
+ */
 @FunctionalInterface
-interface IOConsumer<T> {
+public interface IOConsumer<T> {
     void accept(T t) throws IOException;
 }
