+++ rawhttp-core/src/main/java/rawhttp/core/RawHttpResponse.java
@@ -3,6 +3,7 @@ package rawhttp.core;
 import java.io.IOException;
 import java.util.Optional;
 import javax.annotation.Nullable;
+import rawhttp.core.body.BodyReader;
 import rawhttp.core.body.HttpMessageBody;
 
 /**
