+++ rawhttp-core/src/test/kotlin/rawhttp/core/LazyBodyReaderTest.kt
@@ -4,10 +4,11 @@ import io.kotlintest.matchers.should
 import io.kotlintest.matchers.shouldBe
 import io.kotlintest.matchers.shouldEqual
 import io.kotlintest.specs.StringSpec
-import rawhttp.core.BodyType.CloseTerminated
-import rawhttp.core.BodyType.ContentLength
-import rawhttp.core.BodyType.Encoded
 import rawhttp.core.RawHttpHeaders.Builder.emptyRawHttpHeaders
+import rawhttp.core.body.BodyType.CloseTerminated
+import rawhttp.core.body.BodyType.ContentLength
+import rawhttp.core.body.BodyType.Encoded
+import rawhttp.core.body.LazyBodyReader
 import java.io.ByteArrayOutputStream
 import kotlin.text.Charsets.UTF_8
